Splitting up cdriso.cc a bit.
