+++ src/core/cdriso.cc
@@ -17,33 +17,11 @@
  *   51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.           *
  ***************************************************************************/
 
-#ifdef _WIN32
-#include "support/windowswrapper.h"
-#define strcasecmp _stricmp
-#else
-#include <limits.h>
-#include <sys/time.h>
-#include <unistd.h>
-#endif
 #include <zlib.h>
-
-#include <algorithm>
-#include <filesystem>
 
 #include "core/cdriso.h"
 #include "core/cdrom.h"
 #include "core/iec-60908b-math.h"
-#include "core/ppf.h"
-#include "core/psxemulator.h"
-
-extern "C" {
-#include <libavcodec/avcodec.h>
-#include <libavformat/avformat.h>
-#include <libavutil/mathematics.h>
-#include <libavutil/opt.h>
-#include <libavutil/timestamp.h>
-#include <libswresample/swresample.h>
-}
 
 ////////////////////////////////////////////////////////////////////////////////
 //
@@ -99,934 +77,6 @@ PCSX::CDRiso::trackinfo::cddatype_t PCSX::CDRiso::get_cdda_type(const char *str)
     return trackinfo::BIN;  // no valid extension or no support; assume bin
 }
 
-/* ffmpeg-related code */
-
-static int get_compressed_cdda_track_length(const char *filepath) {
-    int seconds = -1;
-    av_log_set_level(AV_LOG_QUIET);
-
-    AVFormatContext *inAudioFormat = NULL;
-    inAudioFormat = avformat_alloc_context();
-    int errorCode = avformat_open_input(&inAudioFormat, filepath, NULL, NULL);
-    avformat_find_stream_info(inAudioFormat, NULL);
-    seconds = (int)ceil((double)inAudioFormat->duration / (double)AV_TIME_BASE);
-    avformat_close_input(&inAudioFormat);
-    return seconds;
-}
-
-static int decode_packet(int *got_frame, AVPacket pkt, int audio_stream_idx, AVFrame *frame,
-                         AVCodecContext *audio_dec_ctx, void *buf, int *size, SwrContext *swr) {
-    int ret = 0;
-    int decoded = pkt.size;
-    *got_frame = 0;
-
-    if (pkt.stream_index == audio_stream_idx) {
-        // ret = avcodec_decode_audio4(audio_dec_ctx, frame, got_frame, &pkt);
-        ret = avcodec_receive_frame(audio_dec_ctx, frame);
-        if (ret == 0) *got_frame = 1;
-        if (ret == AVERROR(EAGAIN)) ret = 0;
-        if (ret == 0) ret = avcodec_send_packet(audio_dec_ctx, &pkt);
-        if (ret == AVERROR(EAGAIN)) {
-            ret = 0;
-        } else if (ret < 0) {
-            PCSX::g_system->printf(_("Error decoding audio frame\n"));
-            return ret;
-        } else {
-            ret = pkt.size;
-        }
-
-        /* Some audio decoders decode only part of the packet, and have to be
-         * called again with the remainder of the packet data.
-         * Sample: fate-suite/lossless-audio/luckynight-partial.shn
-         * Also, some decoders might over-read the packet. */
-
-        decoded = FFMIN(ret, pkt.size);
-
-        if (*got_frame) {
-            size_t unpadded_linesize = frame->nb_samples * av_get_bytes_per_sample(AVSampleFormat(frame->format));
-            swr_convert(swr, (uint8_t **)&buf, frame->nb_samples, (const uint8_t **)frame->data, frame->nb_samples);
-            (*size) += (unpadded_linesize * 2);
-        }
-    }
-    return decoded;
-}
-
-static int open_codec_context(int *stream_idx, AVFormatContext *fmt_ctx, enum AVMediaType type) {
-    int ret, stream_index;
-    AVStream *st;
-    AVDictionary *opts = NULL;
-
-    ret = av_find_best_stream(fmt_ctx, type, -1, -1, NULL, 0);
-
-    if (ret < 0) {
-        PCSX::g_system->printf(_("Could not find %s stream in input file\n"), av_get_media_type_string(type));
-        return ret;
-    } else {
-        stream_index = ret;
-        st = fmt_ctx->streams[stream_index];
-
-        const AVCodec *dec = avcodec_find_decoder(st->codecpar->codec_id);
-        if (!dec) {
-            PCSX::g_system->printf(_("Failed to find %s codec\n"), av_get_media_type_string(type));
-            return AVERROR(EINVAL);
-        }
-
-        AVCodecContext *dec_ctx = avcodec_alloc_context3(dec);
-        if (!dec_ctx) {
-            PCSX::g_system->printf(_("Failed to find %s codec\n"), av_get_media_type_string(type));
-            return AVERROR(EINVAL);
-        }
-        avcodec_parameters_to_context(dec_ctx, st->codecpar);
-
-        /* Init the decoders, with or without reference counting */
-        if ((ret = avcodec_open2(dec_ctx, dec, NULL)) < 0) {
-            PCSX::g_system->printf(_("Failed to open %s codec\n"), av_get_media_type_string(type));
-            avcodec_free_context(&dec_ctx);
-            return ret;
-        }
-        avcodec_free_context(&dec_ctx);
-        *stream_idx = stream_index;
-    }
-    return 0;
-}
-
-static int decode_compressed_cdda_track(char *buf, char *src_filename, int *size) {
-    AVFormatContext *fmt_ctx = NULL;
-    AVCodecContext *audio_dec_ctx = NULL;
-    const AVCodec *audio_codec = NULL;
-    AVStream *audio_stream = NULL;
-    int audio_stream_idx = -1;
-    AVFrame *frame = NULL;
-    AVPacket pkt;
-    SwrContext *resample_context;
-    int ret = 0, got_frame;
-
-    if (avformat_open_input(&fmt_ctx, src_filename, NULL, NULL) < 0) {
-        PCSX::g_system->printf(_("Could not open source file %s\n"), src_filename);
-        return -1;
-    }
-
-    if (avformat_find_stream_info(fmt_ctx, NULL) < 0) {
-        PCSX::g_system->printf(_("Could not find stream information\n"));
-        ret = -1;
-        goto end;
-    }
-
-    if (open_codec_context(&audio_stream_idx, fmt_ctx, AVMEDIA_TYPE_AUDIO) >= 0) {
-        audio_stream = fmt_ctx->streams[audio_stream_idx];
-    }
-
-    if (!audio_stream) {
-        PCSX::g_system->printf(_("Could not find audio stream in the input, aborting\n"));
-        ret = -1;
-        goto end;
-    }
-
-    audio_codec = avcodec_find_decoder(audio_stream->codecpar->codec_id);
-
-    if (!audio_codec) {
-        PCSX::g_system->printf(_("Could not find audio codec for the input, aborting\n"));
-        ret = -1;
-        goto end;
-    }
-
-    audio_dec_ctx = avcodec_alloc_context3(audio_codec);
-
-    if (!audio_dec_ctx) {
-        PCSX::g_system->printf(_("Could not allocate audio codec for the input, aborting\n"));
-        ret = -1;
-        goto end;
-    }
-
-    // init and configure resampler
-    resample_context = swr_alloc();
-    if (!resample_context) {
-        PCSX::g_system->printf(_("Could not allocate resample context"));
-        ret = -1;
-        goto end;
-    }
-    av_opt_set_int(resample_context, "in_channel_layout", audio_dec_ctx->channel_layout, 0);
-    av_opt_set_int(resample_context, "out_channel_layout", AV_CH_LAYOUT_STEREO, 0);
-    av_opt_set_int(resample_context, "in_sample_rate", audio_dec_ctx->sample_rate, 0);
-    av_opt_set_int(resample_context, "out_sample_rate", 44100, 0);
-    av_opt_set_sample_fmt(resample_context, "in_sample_fmt", audio_dec_ctx->sample_fmt, 0);
-    av_opt_set_sample_fmt(resample_context, "out_sample_fmt", AV_SAMPLE_FMT_S16, 0);
-    if (swr_init(resample_context) < 0) {
-        PCSX::g_system->printf(_("Could not open resample context"));
-        ret = -1;
-        goto end;
-    }
-
-    frame = av_frame_alloc();
-    if (!frame) {
-        PCSX::g_system->printf(_("Could not allocate frame\n"));
-        ret = AVERROR(ENOMEM);
-        goto end;
-    }
-
-    /* initialize packet, set data to NULL, let the demuxer fill it */
-    av_init_packet(&pkt);
-    pkt.data = NULL;
-    pkt.size = 0;
-
-    /* read frames from the file */
-    while (av_read_frame(fmt_ctx, &pkt) >= 0) {
-        AVPacket orig_pkt = pkt;
-        do {
-            ret = decode_packet(&got_frame, pkt, audio_stream_idx, frame, audio_dec_ctx, buf + (*size), size,
-                                resample_context);
-            if (ret < 0) break;
-            pkt.data += ret;
-            pkt.size -= ret;
-        } while (pkt.size > 0);
-        av_packet_unref(&orig_pkt);
-    }
-
-    /* flush cached frames */
-    pkt.data = NULL;
-    pkt.size = 0;
-    do {
-        decode_packet(&got_frame, pkt, audio_stream_idx, frame, audio_dec_ctx, buf + (*size), size, resample_context);
-    } while (got_frame);
-
-end:
-    swr_free(&resample_context);
-    if (audio_dec_ctx) {
-        avcodec_close(audio_dec_ctx);
-        avcodec_free_context(&audio_dec_ctx);
-    }
-    avformat_close_input(&fmt_ctx);
-    av_frame_free(&frame);
-    return ret < 0;
-}
-
-/* end of ffmpeg-only code */
-
-int PCSX::CDRiso::do_decode_cdda(struct trackinfo *tri, uint32_t tracknumber) {
-    tri->decoded_buffer = (char *)malloc(tri->len_decoded_buffer);
-    memset(tri->decoded_buffer, 0, tri->len_decoded_buffer - 1);
-
-    if (tri->decoded_buffer == NULL) {
-        PCSX::g_system->message(_("Could not allocate memory to decode CDDA TRACK: %s\n"), tri->filepath);
-        tri->handle->close();                   // encoded file handle not needed anymore
-        tri->handle.setFile(new BufferFile());  // change handle to decoded one
-        tri->cddatype = trackinfo::BIN;
-        return 0;
-    }
-
-    tri->handle->close();  // encoded file handle not needed anymore
-    tri->handle.reset();
-
-    int ret;
-    PCSX::g_system->printf(_("Decoding audio tr#%u (%s)..."), tracknumber, tri->filepath);
-
-    int len = 0;
-
-    if ((ret = decode_compressed_cdda_track(tri->decoded_buffer, tri->filepath, &len)) == 0) {
-        if (len > tri->len_decoded_buffer) {
-            PCSX::g_system->printf(_("Buffer overflow..."));
-            PCSX::g_system->printf(_("Actual %i vs. %i estimated\n"), len, tri->len_decoded_buffer);
-            len = tri->len_decoded_buffer;  // we probably segfaulted already, oh well...
-        }
-
-        tri->handle.setFile(new BufferFile(tri->decoded_buffer, len));  // change handle to decoded one
-        PCSX::g_system->printf(_("OK\n"), tri->filepath);
-    }
-    tri->cddatype = trackinfo::BIN;
-    return len;
-}
-
-// this function tries to get the .toc file of the given .bin
-// the necessary data is put into the ti (trackinformation)-array
-int PCSX::CDRiso::parsetoc(const char *isofileStr) {
-    std::filesystem::path isofile = MAKEU8(isofileStr);
-    std::filesystem::path tocname, filename;
-    IO<File> fi;
-    char linebuf[256], tmp[256], name[256];
-    char *token;
-    char time[20], time2[20];
-    unsigned int t, sector_offs, sector_size;
-    unsigned int current_zero_gap = 0;
-
-    m_numtracks = 0;
-
-    // copy name of the iso and change extension from .bin to .toc
-    tocname = isofile;
-    tocname.replace_extension("toc");
-
-    fi.setFile(new UvFile(tocname));
-    if (g_emulator->settings.get<Emulator::SettingFullCaching>()) {
-        fi.asA<UvFile>()->startCaching();
-    }
-    if (fi->failed()) {
-        // try changing extension to .cue (to satisfy some stupid tutorials)
-        tocname.replace_extension("cue");
-        fi.setFile(new UvFile(tocname));
-        if (g_emulator->settings.get<Emulator::SettingFullCaching>()) {
-            fi.asA<UvFile>()->startCaching();
-        }
-        if (fi->failed()) {
-            // if filename is image.toc.bin, try removing .bin (for Brasero)
-            tocname = isofile;
-            tocname.replace_extension("");
-            if (tocname.extension() == ".toc") {
-                fi.setFile(new UvFile(tocname));
-                if (g_emulator->settings.get<Emulator::SettingFullCaching>()) {
-                    fi.asA<UvFile>()->startCaching();
-                }
-                if (fi->failed()) {
-                    return -1;
-                }
-            }
-            return -1;
-        }
-    }
-
-    filename = tocname.parent_path();
-
-    memset(&m_ti, 0, sizeof(m_ti));
-    m_cddaBigEndian = true;  // cdrdao uses big-endian for CD Audio
-
-    sector_size = PCSX::CDRom::CD_FRAMESIZE_RAW;
-    sector_offs = 2 * 75;
-
-    // parse the .toc file
-    while (fi->gets(linebuf, sizeof(linebuf))) {
-        // search for tracks
-        strncpy(tmp, linebuf, sizeof(linebuf));
-        token = strtok(tmp, " ");
-
-        if (token == NULL) continue;
-
-        if (!strcmp(token, "TRACK")) {
-            sector_offs += current_zero_gap;
-            current_zero_gap = 0;
-
-            // get type of track
-            token = strtok(NULL, " ");
-            m_numtracks++;
-
-            if (!strncmp(token, "MODE2_RAW", 9)) {
-                m_ti[m_numtracks].type = trackinfo::DATA;
-                // assume data track on 0:2:0
-                m_ti[m_numtracks].start = IEC60908b::MSF(0, 2, 0);
-
-                // check if this image contains mixed subchannel data
-                token = strtok(NULL, " ");
-                if (token != NULL && !strncmp(token, "RW", 2)) {
-                    sector_size = PCSX::CDRom::CD_FRAMESIZE_RAW + PCSX::CDRom::SUB_FRAMESIZE;
-                    m_subChanMixed = true;
-                    if (!strncmp(token, "RW_RAW", 6)) m_subChanRaw = true;
-                }
-            } else if (!strncmp(token, "AUDIO", 5)) {
-                m_ti[m_numtracks].type = trackinfo::CDDA;
-            }
-        } else if (!strcmp(token, "DATAFILE")) {
-            if (m_ti[m_numtracks].type == trackinfo::CDDA) {
-                sscanf(linebuf, "DATAFILE \"%[^\"]\" #%d %8s", name, &t, time2);
-                m_ti[m_numtracks].start_offset = t;
-                t = t / sector_size + sector_offs;
-                m_ti[m_numtracks].start = IEC60908b::MSF(t);
-                m_ti[m_numtracks].length = IEC60908b::MSF(time2);
-            } else {
-                sscanf(linebuf, "DATAFILE \"%[^\"]\" %8s", name, time);
-                m_ti[m_numtracks].length = IEC60908b::MSF(time);
-                m_ti[m_numtracks].handle.setFile(new UvFile(filename / name));
-                if (g_emulator->settings.get<Emulator::SettingFullCaching>()) {
-                    m_ti[m_numtracks].handle.asA<UvFile>()->startCaching();
-                }
-            }
-        } else if (!strcmp(token, "FILE")) {
-            sscanf(linebuf, "FILE \"%[^\"]\" #%d %8s %8s", name, &t, time, time2);
-            m_ti[m_numtracks].start = IEC60908b::MSF(time);
-            t += m_ti[m_numtracks].start.toLBA() * sector_size;
-            m_ti[m_numtracks].start_offset = t;
-            t = t / sector_size + sector_offs;
-            m_ti[m_numtracks].start = IEC60908b::MSF(t);
-            m_ti[m_numtracks].length = IEC60908b::MSF(time2);
-        } else if (!strcmp(token, "ZERO") || !strcmp(token, "SILENCE")) {
-            // skip unneeded optional fields
-            while (token != NULL) {
-                token = strtok(NULL, " ");
-                if (strchr(token, ':') != NULL) break;
-            }
-            if (token != NULL) {
-                current_zero_gap = IEC60908b::MSF(token).toLBA();
-            }
-            if (m_numtracks > 1) {
-                t = m_ti[m_numtracks - 1].start_offset;
-                t /= sector_size;
-                m_pregapOffset = t + m_ti[m_numtracks - 1].length.toLBA();
-            }
-        } else if (!strcmp(token, "START")) {
-            token = strtok(NULL, " ");
-            if (token != NULL && strchr(token, ':')) {
-                t = IEC60908b::MSF(token).toLBA();
-                m_ti[m_numtracks].start_offset += (t - current_zero_gap) * sector_size;
-                t += m_ti[m_numtracks].start.toLBA();
-                m_ti[m_numtracks].start = IEC60908b::MSF(t);
-            }
-        }
-    }
-    if (m_numtracks > 0) m_cdHandle.setFile(new SubFile(m_ti[1].handle, 0, m_ti[1].handle->size()));
-
-    return 0;
-}
-
-// this function tries to get the .cue file of the given .bin
-// the necessary data is put into the ti (trackinformation)-array
-int PCSX::CDRiso::parsecue(const char *isofileString) {
-    std::filesystem::path isofile = MAKEU8(isofileString);
-    std::filesystem::path cuename, filepath;
-    IO<File> fi;
-    char *token;
-    char time[20];
-    char *tmp;
-    char linebuf[256], tmpb[256], dummy[256];
-    unsigned int t, file_len, mode, sector_offs;
-    unsigned int sector_size = 2352;
-
-    m_numtracks = 0;
-
-    // copy name of the iso and change extension from .bin to .cue
-    cuename = isofile;
-    cuename.replace_extension("cue");
-
-    fi.setFile(new UvFile(cuename));
-    if (g_emulator->settings.get<Emulator::SettingFullCaching>()) {
-        fi.asA<UvFile>()->startCaching();
-    }
-    if (fi->failed()) {
-        return -1;
-    }
-
-    // Some stupid tutorials wrongly tell users to use cdrdao to rip a
-    // "bin/cue" image, which is in fact a "bin/toc" image. So let's check
-    // that...
-    if (fi->gets(linebuf, sizeof(linebuf))) {
-        if (!strncmp(linebuf, "CD_ROM_XA", 9)) {
-            // Don't proceed further, as this is actually a .toc file rather
-            // than a .cue file.
-            return parsetoc(isofileString);
-        }
-        fi->rSeek(0, SEEK_SET);
-    }
-
-    // build a path for files referenced in .cue
-    filepath = cuename.parent_path();
-
-    memset(&m_ti, 0, sizeof(m_ti));
-
-    file_len = 0;
-    sector_offs = 2 * 75;
-
-    while (fi->gets(linebuf, sizeof(linebuf))) {
-        strncpy(dummy, linebuf, sizeof(linebuf));
-        token = strtok(dummy, " ");
-
-        if (token == NULL) {
-            continue;
-        }
-
-        if (!strcmp(token, "TRACK")) {
-            m_numtracks++;
-
-            sector_size = 0;
-            if (strstr(linebuf, "AUDIO") != NULL) {
-                m_ti[m_numtracks].type = trackinfo::CDDA;
-                sector_size = PCSX::CDRom::CD_FRAMESIZE_RAW;
-                // Check if extension is mp3, etc, for compressed audio formats
-                if (m_multifile &&
-                    (m_ti[m_numtracks].cddatype = get_cdda_type(m_ti[m_numtracks].filepath)) > trackinfo::BIN) {
-                    int seconds = get_compressed_cdda_track_length(m_ti[m_numtracks].filepath) + 0;
-                    const bool lazy_decode = true;  // TODO: config param
-
-                    // TODO: get frame length for compressed audio as well
-                    m_ti[m_numtracks].len_decoded_buffer = 44100 * (16 / 8) * 2 * seconds;
-                    file_len = m_ti[m_numtracks].len_decoded_buffer / PCSX::CDRom::CD_FRAMESIZE_RAW;
-
-                    // Send to decoder if not lazy decoding
-                    if (!lazy_decode) {
-                        PCSX::g_system->printf("\n");
-                        file_len = do_decode_cdda(&(m_ti[m_numtracks]), m_numtracks) / PCSX::CDRom::CD_FRAMESIZE_RAW;
-                    }
-                }
-            } else if (sscanf(linebuf, " TRACK %u MODE%u/%u", &t, &mode, &sector_size) == 3) {
-                int32_t accurate_len;
-                // TODO: if 2048 frame length -> recalculate file_len?
-                m_ti[m_numtracks].type = trackinfo::DATA;
-                // detect if ECM or compressed & get accurate length
-                if (handleecm(m_ti[m_numtracks].filepath, m_cdHandle, &accurate_len) == 0) {
-                    file_len = accurate_len;
-                }
-            } else {
-                PCSX::g_system->printf(".cue: failed to parse TRACK\n");
-                m_ti[m_numtracks].type = m_numtracks == 1 ? trackinfo::DATA : trackinfo::CDDA;
-            }
-            if (sector_size == 0)  // TODO m_isMode1ISO?
-                sector_size = PCSX::CDRom::CD_FRAMESIZE_RAW;
-        } else if (!strcmp(token, "INDEX")) {
-            if (sscanf(linebuf, " INDEX %02d %8s", &t, time) != 2)
-                PCSX::g_system->printf(".cue: failed to parse INDEX\n");
-            m_ti[m_numtracks].start = IEC60908b::MSF(time);
-
-            t = m_ti[m_numtracks].start.toLBA();
-            m_ti[m_numtracks].start_offset = t * sector_size;
-            t += sector_offs;
-            m_ti[m_numtracks].start = IEC60908b::MSF(t);
-
-            // default track length to file length
-            t = file_len - m_ti[m_numtracks].start_offset / sector_size;
-            m_ti[m_numtracks].length = IEC60908b::MSF(t);
-
-            if (m_numtracks > 1 && !m_ti[m_numtracks].handle) {
-                // this track uses the same file as the last,
-                // start of this track is last track's end
-                t = m_ti[m_numtracks].start.toLBA() - m_ti[m_numtracks - 1].start.toLBA();
-                m_ti[m_numtracks - 1].length = IEC60908b::MSF(t);
-            }
-            if (m_numtracks > 1 && m_pregapOffset == -1) m_pregapOffset = m_ti[m_numtracks].start_offset / sector_size;
-        } else if (!strcmp(token, "PREGAP")) {
-            if (sscanf(linebuf, " PREGAP %8s", time) == 1) {
-                sector_offs += IEC60908b::MSF(time).toLBA();
-            }
-            m_pregapOffset = -1;  // mark to fill track start_offset
-        } else if (!strcmp(token, "FILE")) {
-            t = sscanf(linebuf, " FILE \"%255[^\"]\"", tmpb);
-            if (t != 1) sscanf(linebuf, " FILE %255s", tmpb);
-
-            // absolute path?
-            m_ti[m_numtracks + 1].handle.setFile(new UvFile(tmpb));
-            if (g_emulator->settings.get<Emulator::SettingFullCaching>()) {
-                m_ti[m_numtracks + 1].handle.asA<UvFile>()->startCaching();
-            }
-            if (m_ti[m_numtracks + 1].handle->failed()) {
-                m_ti[m_numtracks + 1].handle.setFile(new UvFile(filepath / tmpb));
-                if (g_emulator->settings.get<Emulator::SettingFullCaching>()) {
-                    m_ti[m_numtracks + 1].handle.asA<UvFile>()->startCaching();
-                }
-            }
-
-            strcpy(m_ti[m_numtracks + 1].filepath,
-                   reinterpret_cast<const char *>(m_ti[m_numtracks + 1].handle->filename().u8string().c_str()));
-
-            // update global offset if this is not first file in this .cue
-            if (m_numtracks + 1 > 1) {
-                m_multifile = true;
-                sector_offs += file_len;
-            }
-
-            file_len = 0;
-            if (m_ti[m_numtracks + 1].handle->failed()) {
-                PCSX::g_system->message(_("\ncould not open: %s\n"), m_ti[m_numtracks + 1].handle->filename().string());
-                m_ti[m_numtracks + 1].handle.reset();
-                continue;
-            }
-
-            // File length, compressed audio length will be calculated in AUDIO tag
-            m_ti[m_numtracks + 1].handle->rSeek(0, SEEK_END);
-            file_len = m_ti[m_numtracks + 1].handle->rTell() / PCSX::CDRom::CD_FRAMESIZE_RAW;
-
-            if (m_numtracks == 0 && (isofile.extension() == ".cue")) {
-                // user selected .cue as image file, use its data track instead
-                m_cdHandle.setFile(new SubFile(m_ti[m_numtracks + 1].handle, 0, m_ti[m_numtracks + 1].handle->size()));
-            }
-        }
-    }
-
-    return 0;
-}
-
-// this function tries to get the .ccd file of the given .img
-// the necessary data is put into the ti (trackinformation)-array
-int PCSX::CDRiso::parseccd(const char *isofileString) {
-    std::filesystem::path ccdname, isofile = MAKEU8(isofileString);
-    IO<File> fi;
-    char linebuf[256];
-    unsigned int t;
-
-    m_numtracks = 0;
-
-    // copy name of the iso and change extension from .img to .ccd
-    ccdname = isofile;
-    ccdname.replace_extension("ccd");
-
-    fi.setFile(new UvFile(ccdname));
-    if (g_emulator->settings.get<Emulator::SettingFullCaching>()) {
-        fi.asA<UvFile>()->startCaching();
-    }
-    if (fi->failed()) {
-        return -1;
-    }
-
-    memset(&m_ti, 0, sizeof(m_ti));
-
-    while (fi->gets(linebuf, sizeof(linebuf))) {
-        if (!strncmp(linebuf, "[TRACK", 6)) {
-            m_numtracks++;
-        } else if (!strncmp(linebuf, "MODE=", 5)) {
-            sscanf(linebuf, "MODE=%d", &t);
-            m_ti[m_numtracks].type = ((t == 0) ? trackinfo::CDDA : trackinfo::DATA);
-        } else if (!strncmp(linebuf, "INDEX 1=", 8)) {
-            sscanf(linebuf, "INDEX 1=%d", &t);
-            m_ti[m_numtracks].start = IEC60908b::MSF(t + 150);
-            m_ti[m_numtracks].start_offset = t * 2352;
-
-            // If we've already seen another track, this is its end
-            if (m_numtracks > 1) {
-                t = m_ti[m_numtracks].start.toLBA() - m_ti[m_numtracks - 1].start.toLBA();
-                m_ti[m_numtracks - 1].length = IEC60908b::MSF(t);
-            }
-        }
-    }
-
-    // Fill out the last track's end based on size
-    if (m_numtracks >= 1) {
-        m_cdHandle->rSeek(0, SEEK_END);
-        t = m_cdHandle->rTell() / PCSX::CDRom::CD_FRAMESIZE_RAW - m_ti[m_numtracks].start.toLBA() + 150;
-        m_ti[m_numtracks].length = IEC60908b::MSF(t);
-    }
-
-    return 0;
-}
-
-// this function tries to get the .mds file of the given .mdf
-// the necessary data is put into the ti (trackinformation)-array
-int PCSX::CDRiso::parsemds(const char *isofileString) {
-    std::filesystem::path mdsname, isofile = MAKEU8(isofileString);
-    IO<File> fi;
-    unsigned int offset, extra_offset, l, i;
-    unsigned short s;
-
-    m_numtracks = 0;
-
-    // copy name of the iso and change extension from .mdf to .mds
-    mdsname = isofile;
-    isofile.replace_extension("mds");
-
-    fi.setFile(new UvFile(mdsname));
-    if (g_emulator->settings.get<Emulator::SettingFullCaching>()) {
-        fi.asA<UvFile>()->startCaching();
-    }
-    if (fi->failed()) return -1;
-
-    memset(&m_ti, 0, sizeof(m_ti));
-
-    // check if it's a valid mds file
-    i = fi->read<uint32_t>();
-    if (i != 0x4944454d) {
-        // not an valid mds file
-        return -1;
-    }
-
-    // get offset to session block
-    fi->rSeek(0x50, SEEK_SET);
-    offset = fi->read<uint32_t>();
-
-    // get total number of tracks
-    offset += 14;
-    fi->rSeek(offset, SEEK_SET);
-    s = fi->read<uint16_t>();
-    m_numtracks = s;
-
-    // get offset to track blocks
-    fi->rSeek(4, SEEK_CUR);
-    offset = fi->read<uint32_t>();
-
-    // skip lead-in data
-    while (1) {
-        fi->rSeek(offset + 4, SEEK_SET);
-        if (fi->getc() < 0xa0) break;
-        offset += 0x50;
-    }
-
-    // check if the image contains mixed subchannel data
-    fi->rSeek(offset + 1, SEEK_SET);
-    m_subChanMixed = m_subChanRaw = (fi->getc() ? true : false);
-
-    // read track data
-    for (i = 1; i <= m_numtracks; i++) {
-        fi->rSeek(offset, SEEK_SET);
-
-        // get the track type
-        m_ti[i].type = ((fi->getc() == 0xA9) ? trackinfo::CDDA : trackinfo::DATA);
-        fi->rSeek(8, SEEK_CUR);
-
-        // get the track starting point
-        m_ti[i].start.m = fi->getc();
-        m_ti[i].start.s = fi->getc();
-        m_ti[i].start.f = fi->getc();
-
-        extra_offset = fi->read<uint32_t>();
-
-        // get track start offset (in .mdf)
-        fi->rSeek(offset + 0x28, SEEK_SET);
-        l = fi->read<uint32_t>();
-        m_ti[i].start_offset = l;
-
-        // get pregap
-        fi->rSeek(extra_offset, SEEK_SET);
-        l = fi->read<uint32_t>();
-        if (l != 0 && i > 1) m_pregapOffset = m_ti[i].start.toLBA();
-
-        // get the track length
-        l = fi->read<uint32_t>();
-        m_ti[i].length = IEC60908b::MSF(l);
-
-        offset += 0x50;
-    }
-
-    return 0;
-}
-
-int PCSX::CDRiso::handlepbp(const char *isofile) {
-    struct {
-        unsigned int sig;
-        unsigned int dontcare[8];
-        unsigned int psar_offs;
-    } pbp_hdr;
-    struct {
-        unsigned char type;
-        unsigned char pad0;
-        unsigned char track;
-        char index0[3];
-        char pad1;
-        char index1[3];
-    } toc_entry;
-    struct {
-        unsigned int offset;
-        unsigned int size;
-        unsigned int dontcare[6];
-    } index_entry;
-    char psar_sig[11];
-    unsigned int t, cd_length, cdimg_base;
-    unsigned int offsettab[8], psisoimg_offs;
-    const char *ext = NULL;
-    int i, ret;
-
-    if (strlen(isofile) >= 4) ext = isofile + strlen(isofile) - 4;
-    if (ext == NULL || (strcmp(ext, ".pbp") != 0 && strcmp(ext, ".PBP") != 0)) return -1;
-
-    m_cdHandle->rSeek(0, SEEK_SET);
-
-    m_numtracks = 0;
-
-    ret = m_cdHandle->read(&pbp_hdr, sizeof(pbp_hdr));
-    if (ret != sizeof(pbp_hdr)) {
-        PCSX::g_system->printf("failed to read pbp\n");
-        goto fail_io;
-    }
-
-    ret = m_cdHandle->rSeek(pbp_hdr.psar_offs, SEEK_SET);
-    if (ret != 0) {
-        PCSX::g_system->printf("failed to seek to %x\n", pbp_hdr.psar_offs);
-        goto fail_io;
-    }
-
-    psisoimg_offs = pbp_hdr.psar_offs;
-    m_cdHandle->read(psar_sig, sizeof(psar_sig));
-    psar_sig[10] = 0;
-    if (strcmp(psar_sig, "PSTITLEIMG") == 0) {
-        // multidisk image?
-        ret = m_cdHandle->rSeek(pbp_hdr.psar_offs + 0x200, SEEK_SET);
-        if (ret != 0) {
-            PCSX::g_system->printf("failed to seek to %x\n", pbp_hdr.psar_offs + 0x200);
-            goto fail_io;
-        }
-
-        if (m_cdHandle->read(&offsettab, sizeof(offsettab)) != sizeof(offsettab)) {
-            PCSX::g_system->printf("failed to read offsettab\n");
-            goto fail_io;
-        }
-
-        for (i = 0; i < sizeof(offsettab) / sizeof(offsettab[0]); i++) {
-            if (offsettab[i] == 0) break;
-        }
-        m_cdrIsoMultidiskCount = i;
-        if (m_cdrIsoMultidiskCount == 0) {
-            PCSX::g_system->printf("multidisk eboot has 0 images?\n");
-            goto fail_io;
-        }
-
-        if (m_cdrIsoMultidiskSelect >= m_cdrIsoMultidiskCount) m_cdrIsoMultidiskSelect = 0;
-
-        psisoimg_offs += offsettab[m_cdrIsoMultidiskSelect];
-
-        ret = m_cdHandle->rSeek(psisoimg_offs, SEEK_SET);
-        if (ret != 0) {
-            PCSX::g_system->printf("failed to seek to %x\n", psisoimg_offs);
-            goto fail_io;
-        }
-
-        m_cdHandle->read(psar_sig, sizeof(psar_sig));
-        psar_sig[10] = 0;
-    }
-
-    if (strcmp(psar_sig, "PSISOIMG00") != 0) {
-        PCSX::g_system->printf("bad psar_sig: %s\n", psar_sig);
-        goto fail_io;
-    }
-
-    // seek to TOC
-    ret = m_cdHandle->rSeek(psisoimg_offs + 0x800, SEEK_SET);
-    if (ret != 0) {
-        PCSX::g_system->printf("failed to seek to %x\n", psisoimg_offs + 0x800);
-        goto fail_io;
-    }
-
-    // first 3 entries are special
-    m_cdHandle->rSeek(sizeof(toc_entry), SEEK_CUR);
-    m_cdHandle->read(&toc_entry, sizeof(toc_entry));
-    m_numtracks = PCSX::CDRom::btoi(toc_entry.index1[0]);
-
-    m_cdHandle->read(&toc_entry, sizeof(toc_entry));
-    cd_length = PCSX::CDRom::btoi(toc_entry.index1[0]) * 60 * 75 + PCSX::CDRom::btoi(toc_entry.index1[1]) * 75 +
-                PCSX::CDRom::btoi(toc_entry.index1[2]);
-
-    for (i = 1; i <= m_numtracks; i++) {
-        m_cdHandle->read(&toc_entry, sizeof(toc_entry));
-
-        m_ti[i].type = (toc_entry.type == 1) ? trackinfo::CDDA : trackinfo::DATA;
-
-        m_ti[i].start_offset = PCSX::CDRom::btoi(toc_entry.index0[0]) * 60 * 75 +
-                               PCSX::CDRom::btoi(toc_entry.index0[1]) * 75 + PCSX::CDRom::btoi(toc_entry.index0[2]);
-        m_ti[i].start_offset *= 2352;
-        m_ti[i].start.m = PCSX::CDRom::btoi(toc_entry.index1[0]);
-        m_ti[i].start.s = PCSX::CDRom::btoi(toc_entry.index1[1]);
-        m_ti[i].start.f = PCSX::CDRom::btoi(toc_entry.index1[2]);
-
-        if (i > 1) {
-            t = m_ti[i].start.toLBA() - m_ti[i - 1].start.toLBA();
-            m_ti[i - 1].length = IEC60908b::MSF(t);
-        }
-    }
-    t = cd_length - m_ti[m_numtracks].start_offset / 2352;
-    m_ti[m_numtracks].length = IEC60908b::MSF(t);
-
-    // seek to ISO index
-    ret = m_cdHandle->rSeek(psisoimg_offs + 0x4000, SEEK_SET);
-    if (ret != 0) {
-        PCSX::g_system->printf("failed to seek to ISO index\n");
-        goto fail_io;
-    }
-
-    m_compr_img = (compr_img_t *)calloc(1, sizeof(*m_compr_img));
-    if (m_compr_img == NULL) goto fail_io;
-
-    m_compr_img->block_shift = 4;
-    m_compr_img->current_block = (unsigned int)-1;
-
-    m_compr_img->index_len = (0x100000 - 0x4000) / sizeof(index_entry);
-    m_compr_img->index_table =
-        (unsigned int *)malloc((m_compr_img->index_len + 1) * sizeof(m_compr_img->index_table[0]));
-    if (m_compr_img->index_table == NULL) goto fail_io;
-
-    cdimg_base = psisoimg_offs + 0x100000;
-    for (i = 0; i < m_compr_img->index_len; i++) {
-        ret = m_cdHandle->read(&index_entry, sizeof(index_entry));
-        if (ret != sizeof(index_entry)) {
-            PCSX::g_system->printf("failed to read index_entry #%d\n", i);
-            goto fail_index;
-        }
-
-        if (index_entry.size == 0) break;
-
-        m_compr_img->index_table[i] = cdimg_base + index_entry.offset;
-    }
-    m_compr_img->index_table[i] = cdimg_base + index_entry.offset + index_entry.size;
-
-    return 0;
-
-fail_index:
-    free(m_compr_img->index_table);
-    m_compr_img->index_table = NULL;
-fail_io:
-    if (m_compr_img != NULL) {
-        free(m_compr_img);
-        m_compr_img = NULL;
-    }
-    return -1;
-}
-
-int PCSX::CDRiso::handlecbin(const char *isofile) {
-    struct {
-        char magic[4];
-        unsigned int header_size;
-        unsigned long long total_bytes;
-        unsigned int block_size;
-        unsigned char ver;  // 1
-        unsigned char align;
-        unsigned char rsv_06[2];
-    } ciso_hdr;
-    const char *ext = NULL;
-    unsigned int index = 0, plain;
-    int i, ret;
-    size_t read_len = 0;
-
-    if (strlen(isofile) >= 5) ext = isofile + strlen(isofile) - 5;
-    if (ext == NULL || (strcasecmp(ext + 1, ".cbn") != 0 && strcasecmp(ext, ".cbin") != 0)) return -1;
-
-    m_cdHandle->rSeek(0, SEEK_SET);
-
-    ret = m_cdHandle->read(&ciso_hdr, sizeof(ciso_hdr));
-    if (ret != sizeof(ciso_hdr)) {
-        PCSX::g_system->printf("failed to read ciso header\n");
-        return -1;
-    }
-
-    if (strncmp(ciso_hdr.magic, "CISO", 4) != 0 || ciso_hdr.total_bytes <= 0 || ciso_hdr.block_size <= 0) {
-        PCSX::g_system->printf("bad ciso header\n");
-        return -1;
-    }
-    if (ciso_hdr.header_size != 0 && ciso_hdr.header_size != sizeof(ciso_hdr)) {
-        ret = m_cdHandle->rSeek(ciso_hdr.header_size, SEEK_SET);
-        if (ret != 0) {
-            PCSX::g_system->printf("failed to seek to %x\n", ciso_hdr.header_size);
-            return -1;
-        }
-    }
-
-    m_compr_img = (compr_img_t *)calloc(1, sizeof(*m_compr_img));
-    if (m_compr_img == NULL) goto fail_io;
-
-    m_compr_img->block_shift = 0;
-    m_compr_img->current_block = (unsigned int)-1;
-
-    m_compr_img->index_len = ciso_hdr.total_bytes / ciso_hdr.block_size;
-    m_compr_img->index_table =
-        (unsigned int *)malloc((m_compr_img->index_len + 1) * sizeof(m_compr_img->index_table[0]));
-    if (m_compr_img->index_table == NULL) goto fail_io;
-
-    read_len = sizeof(m_compr_img->index_table[0]) * m_compr_img->index_len;
-    ret = m_cdHandle->read(m_compr_img->index_table, read_len);
-    if (ret != read_len) {
-        PCSX::g_system->printf("failed to read index table\n");
-        goto fail_index;
-    }
-
-    for (i = 0; i < m_compr_img->index_len + 1; i++) {
-        index = m_compr_img->index_table[i];
-        plain = index & 0x80000000;
-        index &= 0x7fffffff;
-        m_compr_img->index_table[i] = (index << ciso_hdr.align) | plain;
-    }
-    if ((int64_t)index << ciso_hdr.align >= 0x80000000ll) {
-        PCSX::g_system->printf("warning: ciso img too large, expect problems\n");
-    }
-
-    return 0;
-
-fail_index:
-    free(m_compr_img->index_table);
-    m_compr_img->index_table = NULL;
-fail_io:
-    if (m_compr_img != NULL) {
-        free(m_compr_img);
-        m_compr_img = NULL;
-    }
-    return -1;
-}
-
 // this function tries to get the .sub file of the given .img
 int PCSX::CDRiso::opensubfile(const char *isoname) {
     char subname[MAXPATHLEN];
@@ -1062,85 +112,6 @@ int PCSX::CDRiso::opensubfile(const char *isoname) {
     }
 
     return 0;
-}
-
-int PCSX::CDRiso::LoadSBI(const char *filename) {
-    IO<File> sbihandle;
-    char buffer[16], sbifile[MAXPATHLEN];
-
-    if (filename == NULL) {
-        if (PCSX::g_emulator->m_cdromId[0] == '\0') return -1;
-
-        // Generate filename in the format of SLUS_123.45.sbi
-        buffer[0] = toupper(PCSX::g_emulator->m_cdromId[0]);
-        buffer[1] = toupper(PCSX::g_emulator->m_cdromId[1]);
-        buffer[2] = toupper(PCSX::g_emulator->m_cdromId[2]);
-        buffer[3] = toupper(PCSX::g_emulator->m_cdromId[3]);
-        buffer[4] = '_';
-        buffer[5] = PCSX::g_emulator->m_cdromId[4];
-        buffer[6] = PCSX::g_emulator->m_cdromId[5];
-        buffer[7] = PCSX::g_emulator->m_cdromId[6];
-        buffer[8] = '.';
-        buffer[9] = PCSX::g_emulator->m_cdromId[7];
-        buffer[10] = PCSX::g_emulator->m_cdromId[8];
-        buffer[11] = '.';
-        buffer[12] = 's';
-        buffer[13] = 'b';
-        buffer[14] = 'i';
-        buffer[15] = '\0';
-
-        sprintf(sbifile, "%s%s", PCSX::g_emulator->settings.get<Emulator::SettingPpfDir>().string().c_str(), buffer);
-        filename = sbifile;
-    }
-
-    sbihandle.setFile(new UvFile(filename));
-    if (g_emulator->settings.get<Emulator::SettingFullCaching>()) {
-        sbihandle.asA<UvFile>()->startCaching();
-    }
-    if (sbihandle->failed()) {
-        return -1;
-    }
-
-    // init
-    sbicount = 0;
-
-    // 4-byte SBI header
-    sbihandle->read(buffer, 4);
-    while (!sbihandle->eof()) {
-        sbihandle->read(sbitime[sbicount++], 3);
-        sbihandle->read(buffer, 11);
-    }
-
-    PCSX::g_system->printf(_("Loaded SBI file: %s.\n"), filename);
-
-    return 0;
-}
-
-bool PCSX::CDRiso::CheckSBI(const uint8_t *time) {
-    int lcv;
-
-    // both BCD format
-    for (lcv = 0; lcv < sbicount; lcv++) {
-        if (time[0] == sbitime[lcv][0] && time[1] == sbitime[lcv][1] && time[2] == sbitime[lcv][2]) return true;
-    }
-
-    return false;
-}
-
-void PCSX::CDRiso::UnloadSBI() { sbicount = 0; }
-
-int PCSX::CDRiso::opensbifile(const char *isoname) {
-    char sbiname[MAXPATHLEN];
-
-    strncpy(sbiname, isoname, sizeof(sbiname));
-    sbiname[MAXPATHLEN - 1] = '\0';
-    if (strlen(sbiname) >= 4) {
-        strcpy(sbiname + strlen(sbiname) - 4, ".sbi");
-    } else {
-        return -1;
-    }
-
-    return LoadSBI(sbiname);
 }
 
 ssize_t PCSX::CDRiso::cdread_normal(IO<File> f, unsigned int base, void *dest, int sector) {
@@ -1268,309 +239,6 @@ ssize_t PCSX::CDRiso::cdread_2048(IO<File> f, unsigned int base, void *dest, int
     return ret;
 }
 
-/* Adapted from ecm.c:unecmify() (C) Neill Corlett */
-ssize_t PCSX::CDRiso::ecmDecode(IO<File> f, unsigned int base, void *dest, int sector) {
-    uint32_t output_edc = 0, b = 0, writebytecount = 0, num;
-    uint32_t sectorcount = 0;
-    int8_t type = 0;  // mode type 0 (META) or 1, 2 or 3 for CDROM type
-    uint8_t sector_buffer[PCSX::CDRom::CD_FRAMESIZE_RAW];
-    // this flag tells if to decode all sectors or just skip to wanted sector
-    bool processsectors = (bool)m_decoded_ecm_sectors;
-    ECMFILELUT *pos = &(m_ecm_savetable[0]);  // points always to beginning of ECM DATA
-
-    // If not pointing to ECM file but CDDA file or some other track
-    if (f != m_cdHandle) {
-        // printf("BASETR %i %i\n", base, sector);
-        return (*this.*m_cdimg_read_func_o)(f, base, dest, sector);
-    }
-    // When sector exists in decoded ECM file buffer
-    else if (m_decoded_ecm_sectors && sector < m_decoded_ecm_sectors) {
-        // printf("ReadSector %i %i\n", sector, savedsectors);
-        return (*this.*m_cdimg_read_func_o)(m_decoded_ecm, base, dest, sector);
-    }
-    // To prevent invalid seek
-    /* else if (sector > m_len_ecm_savetable) {
-            PCSX::g_system->printf("ECM: invalid sector requested\n");
-            return -1;
-    }*/
-    // printf("SeekSector %i %i %i %i\n", sector, pos->sector, m_prevsector, base);
-
-    if (sector <= m_len_ecm_savetable) {
-        // get sector from LUT which points to wanted sector or close to
-        // TODO: What would be optimal maximum to search near sector?
-        //       Might cause slowdown if too small but too big also..
-        for (sectorcount = sector; ((sectorcount > 0) && ((sector - sectorcount) <= 50000)); sectorcount--) {
-            if (m_ecm_savetable[sectorcount].filepos >= ECM_HEADER_SIZE) {
-                pos = &(m_ecm_savetable[sectorcount]);
-                // printf("LUTSector %i %i %i %i\n", sector, pos->sector, m_prevsector, base);
-                break;
-            }
-        }
-        // if suitable sector was not found from LUT use last sector if less than wanted sector
-        if (pos->filepos <= ECM_HEADER_SIZE && sector > m_prevsector) pos = &(m_ecm_savetable[m_prevsector]);
-    }
-
-    auto reconstructSector = [](uint8_t *sector, int8_t type) {
-        auto ref32 = [sector](uint32_t offset) -> uint32_t & { return *reinterpret_cast<uint32_t *>(sector + offset); };
-        // Sync
-        sector[0x000] = 0x00;
-        sector[0x001] = 0xff;
-        sector[0x002] = 0xff;
-        sector[0x003] = 0xff;
-        sector[0x004] = 0xff;
-        sector[0x005] = 0xff;
-        sector[0x006] = 0xff;
-        sector[0x007] = 0xff;
-        sector[0x008] = 0xff;
-        sector[0x009] = 0xff;
-        sector[0x00a] = 0xff;
-        sector[0x00b] = 0x00;
-
-        switch (type) {
-            case 1:
-                // Mode
-                sector[0x00f] = 0x01;
-                // Empty
-                sector[0x814] = 0x00;
-                sector[0x815] = 0x00;
-                sector[0x816] = 0x00;
-                sector[0x817] = 0x00;
-                sector[0x818] = 0x00;
-                sector[0x819] = 0x00;
-                sector[0x81a] = 0x00;
-                sector[0x81b] = 0x00;
-                break;
-            case 2:
-            case 3:
-                // Mode
-                sector[0x00f] = 0x02;
-                // Subheaders
-                sector[0x010] = sector[0x014];
-                sector[0x011] = sector[0x015];
-                sector[0x012] = sector[0x016];
-                sector[0x013] = sector[0x017];
-                break;
-        }
-
-        // Compute EDC
-        switch (type) {
-            case 1:
-                ref32(0x810) = SWAP_LE32(IEC60908b::computeEDC(0, sector, 0x810));
-                break;
-            case 2:
-                ref32(0x818) = SWAP_LE32(IEC60908b::computeEDC(0, sector + 0x10, 0x808));
-                break;
-            case 3:
-                ref32(0x92c) = SWAP_LE32(IEC60908b::computeEDC(0, sector + 0x10, 0x91c));
-                break;
-        }
-
-        // Compute ECC
-        switch (type) {
-            case 1:
-                IEC60908b::computeECC(sector + 0xc, sector + 0x10, sector + 0x81c);
-                break;
-            case 2:
-                IEC60908b::computeECC(ZEROADDRESS, sector + 0x10, sector + 0x81c);
-                break;
-        }
-    };
-
-    writebytecount = pos->sector * PCSX::CDRom::CD_FRAMESIZE_RAW;
-    sectorcount = pos->sector;
-    if (m_decoded_ecm_sectors) m_decoded_ecm->rSeek(writebytecount, SEEK_SET);  // rewind to last pos
-    f->rSeek(/*base+*/ pos->filepos, SEEK_SET);
-    while (sector >= sectorcount) {  // decode ecm file until we are past wanted sector
-        int c = f->getc();
-        int bits = 5;
-        if (c == EOF) {
-            goto error_in;
-        }
-        type = c & 3;
-        num = (c >> 2) & 0x1F;
-        // printf("ECM1 file; count %x\n", c);
-        while (c & 0x80) {
-            c = f->getc();
-            // printf("ECM2 file; count %x\n", c);
-            if (c == EOF) {
-                goto error_in;
-            }
-            if ((bits > 31) || ((uint32_t)(c & 0x7F)) >= (((uint32_t)0x80000000LU) >> (bits - 1))) {
-                // PCSX::g_system->message(_("Corrupt ECM file; invalid sector count\n"));
-                goto error;
-            }
-            num |= ((uint32_t)(c & 0x7F)) << bits;
-            bits += 7;
-        }
-        if (num == 0xFFFFFFFF) {
-            // End indicator
-            m_len_decoded_ecm_buffer = writebytecount;
-            m_len_ecm_savetable = m_len_decoded_ecm_buffer / PCSX::CDRom::CD_FRAMESIZE_RAW;
-            break;
-        }
-        num++;
-        while (num) {
-            if (!processsectors && sectorcount >= (sector - 1)) {  // ensure that we read the sector we are supposed to
-                processsectors = true;
-                // printf("Saving at %i\n", sectorcount);
-            } else if (processsectors && sectorcount > sector) {
-                // printf("Terminating at %i\n", sectorcount);
-                break;
-            }
-            /*printf("Type %i Num %i SeekSector %i ProcessedSectors %i(%i) Bytecount %i Pos %li Write %u\n",
-                            type, num, sector, sectorcount, pos->sector, writebytecount, ftell(f),
-               processsectors);*/
-            switch (type) {
-                case 0:  // META
-                    b = num;
-                    if (b > sizeof(sector_buffer)) {
-                        b = sizeof(sector_buffer);
-                    }
-                    writebytecount += b;
-                    if (!processsectors) {
-                        f->rSeek(b, SEEK_CUR);
-                        break;
-                    }  // seek only
-                    if (f->read(sector_buffer, b) != b) {
-                        goto error_in;
-                    }
-                    // output_edc = edc_compute(output_edc, sector_buffer, b);
-                    if (m_decoded_ecm_sectors &&
-                        m_decoded_ecm->write(sector_buffer, b) != b) {  // just seek or write also
-                        goto error_out;
-                    }
-                    break;
-                case 1:  // Mode 1
-                    b = 1;
-                    writebytecount += ECM_SECTOR_SIZE[type];
-                    if (f->read(sector_buffer + 0x00C, 0x003) != 0x003) {
-                        goto error_in;
-                    }
-                    if (f->read(sector_buffer + 0x010, 0x800) != 0x800) {
-                        goto error_in;
-                    }
-                    if (!processsectors) break;  // seek only
-                    reconstructSector(sector_buffer, type);
-                    // output_edc = edc_compute(output_edc, sector_buffer, ECM_SECTOR_SIZE[type]);
-                    if (m_decoded_ecm_sectors &&
-                        m_decoded_ecm->write(sector_buffer, ECM_SECTOR_SIZE[type]) != ECM_SECTOR_SIZE[type]) {
-                        goto error_out;
-                    }
-                    break;
-                case 2:  // Mode 2 (XA), form 1
-                    b = 1;
-                    writebytecount += ECM_SECTOR_SIZE[type];
-                    if (!processsectors) {
-                        f->rSeek(0x804, SEEK_CUR);
-                        break;
-                    }  // seek only
-                    if (f->read(sector_buffer + 0x014, 0x804) != 0x804) {
-                        goto error_in;
-                    }
-                    reconstructSector(sector_buffer, type);
-                    // output_edc = edc_compute(output_edc, sector_buffer + 0x10, ECM_SECTOR_SIZE[type]);
-                    if (m_decoded_ecm_sectors &&
-                        m_decoded_ecm->write(sector_buffer + 0x10, ECM_SECTOR_SIZE[type]) != ECM_SECTOR_SIZE[type]) {
-                        goto error_out;
-                    }
-                    break;
-                case 3:  // Mode 2 (XA), form 2
-                    b = 1;
-                    writebytecount += ECM_SECTOR_SIZE[type];
-                    if (!processsectors) {
-                        f->rSeek(0x918, SEEK_CUR);
-                        break;
-                    }  // seek only
-                    if (f->read(sector_buffer + 0x014, 0x918) != 0x918) {
-                        goto error_in;
-                    }
-                    reconstructSector(sector_buffer, type);
-                    // output_edc = edc_compute(output_edc, sector_buffer + 0x10, ECM_SECTOR_SIZE[type]);
-                    if (m_decoded_ecm_sectors &&
-                        m_decoded_ecm->write(sector_buffer + 0x10, ECM_SECTOR_SIZE[type]) != ECM_SECTOR_SIZE[type]) {
-                        goto error_out;
-                    }
-                    break;
-            }
-            sectorcount = ((writebytecount / PCSX::CDRom::CD_FRAMESIZE_RAW) - 0);
-            num -= b;
-        }
-        if (type && sectorcount > 0 && m_ecm_savetable[sectorcount].filepos <= ECM_HEADER_SIZE) {
-            m_ecm_savetable[sectorcount].filepos = f->rTell() /*-base*/;
-            m_ecm_savetable[sectorcount].sector = sectorcount;
-            // printf("Marked %i at pos %i\n", m_ecm_savetable[sectorcount].sector,
-            // m_ecm_savetable[sectorcount].filepos);
-        }
-    }
-
-    if (m_decoded_ecm_sectors) {
-        m_decoded_ecm->rSeek(-1 * PCSX::CDRom::CD_FRAMESIZE_RAW, SEEK_CUR);
-        num = m_decoded_ecm->read(sector_buffer, PCSX::CDRom::CD_FRAMESIZE_RAW);
-        m_decoded_ecm_sectors = std::max(m_decoded_ecm_sectors, sectorcount);
-    } else {
-        num = PCSX::CDRom::CD_FRAMESIZE_RAW;
-    }
-
-    memcpy(dest, sector_buffer, PCSX::CDRom::CD_FRAMESIZE_RAW);
-    m_prevsector = sectorcount;
-    // printf("OK: Frame decoded %i %i\n", sectorcount-1, writebytecount);
-    return num;
-
-error_in:
-error:
-error_out:
-    // memset(dest, 0x0, PCSX::CDRomCD_FRAMESIZE_RAW);
-    PCSX::g_system->printf("Error decoding ECM image: WantedSector %i Type %i Base %i Sectors %i(%i) Pos %i(%li)\n",
-                           sector, type, base, sectorcount, pos->sector, writebytecount, f->rTell());
-    return -1;
-}
-
-int PCSX::CDRiso::handleecm(const char *isoname, IO<File> cdh, int32_t *accurate_length) {
-    // Rewind to start and check ECM header and filename suffix validity
-    cdh->rSeek(0, SEEK_SET);
-    if ((cdh->getc() == 'E') && (cdh->getc() == 'C') && (cdh->getc() == 'M') && (cdh->getc() == 0x00) &&
-        (strncmp((isoname + strlen(isoname) - 5), ".ecm", 4))) {
-        // Function used to read CD normally
-        // TODO: detect if 2048 and use it
-        m_cdimg_read_func_o = &CDRiso::cdread_normal;
-
-        // Function used to decode ECM data
-        m_cdimg_read_func = &CDRiso::ecmDecode;
-
-        // Last accessed sector
-        m_prevsector = 0;
-
-        // Already analyzed during this session, use cached results
-        if (m_ecm_file_detected) {
-            if (accurate_length) *accurate_length = m_len_ecm_savetable;
-            return 0;
-        }
-
-        PCSX::g_system->printf(_("\nDetected ECM file with proper header and filename suffix.\n"));
-
-        // Reserve maximum known sector ammount for LUT (80MIN CD)
-        m_len_ecm_savetable = 75 * 80 * 60;  // 2*(accurate_length/PCSX::CDRomCD_FRAMESIZE_RAW);
-
-        // Index 0 always points to beginning of ECM data
-        m_ecm_savetable = (ECMFILELUT *)calloc(m_len_ecm_savetable, sizeof(ECMFILELUT));  // calloc returns nulled data
-        m_ecm_savetable[0].filepos = ECM_HEADER_SIZE;
-
-        if (accurate_length || m_decoded_ecm_sectors) {
-            uint8_t tbuf1[PCSX::CDRom::CD_FRAMESIZE_RAW];
-            m_len_ecm_savetable = 0;             // indicates to cdread_ecm_decode that no lut has been built yet
-            ecmDecode(cdh, 0U, tbuf1, INT_MAX);  // builds LUT completely
-            if (accurate_length) *accurate_length = m_len_ecm_savetable;
-        }
-
-        // Full image decoded? Needs fmemopen()
-
-        m_ecm_file_detected = true;
-
-        return 0;
-    }
-    return -1;
-}
-
 uint8_t *PCSX::CDRiso::getBuffer() {
     if (m_useCompressed) {
         return m_compr_img->buff_raw[m_compr_img->sector_in_blk] + 12;
@@ -1579,7 +247,7 @@ uint8_t *PCSX::CDRiso::getBuffer() {
     }
 }
 
-void PCSX::CDRiso::PrintTracks() {
+void PCSX::CDRiso::printTracks() {
     for (int i = 1; i <= m_numtracks; i++) {
         PCSX::g_system->printf(_("Track %.2d (%s) - Start %.2d:%.2d:%.2d, Length %.2d:%.2d:%.2d\n"), i,
                                (m_ti[i].type == trackinfo::DATA        ? "DATA"
@@ -1649,12 +317,9 @@ bool PCSX::CDRiso::open(void) {
 
     if (!m_ecm_file_detected) {
         // guess whether it is mode1/2048
-        m_cdHandle->rSeek(0, SEEK_END);
-        if (m_cdHandle->rTell() % 2048 == 0) {
-            unsigned int modeTest = 0;
-            m_cdHandle->rSeek(0, SEEK_SET);
-            m_cdHandle->read(&modeTest, 4);
-            if (SWAP_LE32(modeTest) != 0xffffff00) {
+        if (m_cdHandle->size() % 2048 == 0) {
+            unsigned int modeTest = m_cdHandle->readAt<uint32_t>(0);
+            if (modeTest != 0xffffff00) {
                 PCSX::g_system->printf("[2048]");
                 m_isMode1ISO = true;
             }
@@ -1671,7 +336,7 @@ bool PCSX::CDRiso::open(void) {
 
     PCSX::g_system->printf(".\n");
 
-    PrintTracks();
+    printTracks();
 
     if (m_subChanMixed && (m_cdimg_read_func == &CDRiso::cdread_normal)) {
         m_cdimg_read_func = &CDRiso::cdread_sub_mixed;
